requirements.txt
Flask==0.12.2
python-jose==1.4.0
cryptography==2.0.3